+++ lib/src/decoration.dart
@@ -1,6 +1,25 @@
 import 'package:flutter/material.dart';
 
-enum PencilDecorationType { blank, lines, chequered, dots }
+enum PencilDecorationType { blank, lines, chequered, dots, custom }
+
+/// A custom painter in case a special pattern is needed.
+typedef PencilDecorationCustomPainter = void Function({
+// decoration object
+  PencilDecoration decoration,
+
+// The paint based on the decoration
+  Paint paint,
+
+// Pre-calculated coordinates reflecting the potentially defined padding
+  double xStart,
+  double yStart,
+  double width,
+  double height,
+
+// Canvas and size like for every paint function
+  Canvas canvas,
+  Size size,
+});
 
 class PencilDecoration {
   final PencilDecorationType type;
@@ -8,7 +27,11 @@ class PencilDecoration {
   final Color backgroundColor;
   final Color patternColor;
   final double lineWidth;
-  final double numberOfLines;
+  final bool hasFrame;
+  final double? numberOfLines;
+  final double? spacing;
+
+  final PencilDecorationCustomPainter? customPainter;
 
   PencilDecoration({
     this.type = PencilDecorationType.blank,
@@ -16,35 +39,70 @@ class PencilDecoration {
     this.backgroundColor = Colors.transparent,
     this.patternColor = Colors.black54,
     this.lineWidth = 1.0,
-    this.numberOfLines = 0,
+    this.hasFrame = false,
+    this.numberOfLines,
+    this.spacing,
+    this.customPainter,
   }) {
     assert(() {
-      if (type != PencilDecorationType.blank && numberOfLines <= 0) {
+      if (type == PencilDecorationType.lines ||
+          type == PencilDecorationType.chequered ||
+          type == PencilDecorationType.dots) {
+        if (numberOfLines == null && spacing == null) {
+          debugPrint(
+            'Either numberOfLines or spacing is needed.',
+          );
+          return false;
+        }
+        if (numberOfLines != null && spacing != null) {
+          debugPrint(
+            'Either numberOfLines or spacing can be given.',
+          );
+          return false;
+        }
+        if (numberOfLines != null) {
+          if (numberOfLines! <= 0) {
+            debugPrint(
+              'Number of lines must be a positive number.',
+            );
+            return false;
+          }
+        }
+        if (spacing != null) {
+          if (spacing! <= 0) {
+            debugPrint(
+              'Spacing must be larger than 0.',
+            );
+            return false;
+          }
+        }
+      }
+      if (type == PencilDecorationType.custom && customPainter == null) {
         debugPrint(
-          'numberOfLines must be larger than 0 if a patterned '
-          'background shall be used.',
+          'When using a custom decoration a painter callback mut be given.',
         );
         return false;
       }
+
       return true;
     }());
   }
 
+  /// Paint the decoration. This paint is always done before the actual content
+  /// is painted.
   void paint(Canvas canvas, Size size) {
     if (type == PencilDecorationType.blank) return;
 
+    // All predefined pattern go here
     final paint = Paint()
       ..color = patternColor
       ..strokeWidth = lineWidth;
+
+    // Width and height of the pattern
     final patternHeight = size.height - padding.top - padding.bottom;
     final patternWidth = size.width - padding.left - padding.right;
-    final double ySpacing =
-        numberOfLines == 1 ? patternHeight : patternHeight / numberOfLines;
-    final double xSpacing = ySpacing;
-    final double xStart = padding.left;
-    final double yStart = padding.top;
 
-    // there might be padding setting and layout constraints that lead negative
+    // There might be padding setting and layout constraints that lead negative
     // heights or width. In this case nothing is painted
     if (patternHeight <= 0.0 || patternWidth <= 0.0) {
       debugPrint(
@@ -54,6 +112,43 @@ class PencilDecoration {
       );
       return;
     }
+
+    // Starting points of the pattern
+    final double xStart = padding.left;
+    final double yStart = padding.top;
+
+    // If a customer painter is defined it is called and the functions returns.
+    if (type == PencilDecorationType.custom) {
+      customPainter?.call(
+        decoration: this,
+        paint: paint,
+        xStart: xStart,
+        yStart: yStart,
+        width: patternWidth,
+        height: patternHeight,
+        canvas: canvas,
+        size: size,
+      );
+      return;
+    }
+
+    // ySpacing will be calculated either based on number of lines or the
+    // spacing given.
+    late double ySpacing;
+    if (type == PencilDecorationType.lines ||
+        type == PencilDecorationType.chequered ||
+        type == PencilDecorationType.dots) {
+      if (numberOfLines != null) {
+        ySpacing =
+            numberOfLines == 1 ? patternHeight : patternHeight / numberOfLines!;
+      }
+      if (spacing != null) {
+        ySpacing = spacing!;
+      }
+    }
+
+    // x spacing is the same as y spacing for the moment
+    final double xSpacing = ySpacing;
 
     // Draw vertical lines.
     if (type == PencilDecorationType.chequered) {
@@ -69,7 +164,7 @@ class PencilDecoration {
     // Draw horizontal lines
     if (type == PencilDecorationType.lines ||
         type == PencilDecorationType.chequered) {
-      for (int iY = 0; iY < numberOfLines + 1; iY++) {
+      for (int iY = 0; iY < patternHeight / ySpacing; iY++) {
         canvas.drawLine(
           Offset(xStart, yStart + ySpacing * iY),
           Offset(xStart + patternWidth, yStart + ySpacing * iY),
@@ -81,7 +176,7 @@ class PencilDecoration {
     // Draw the dotted pattern
     if (type == PencilDecorationType.dots) {
       for (int iX = 0; iX < patternWidth / xSpacing; iX++) {
-        for (int iY = 0; iY < numberOfLines + 1; iY++) {
+        for (int iY = 0; iY < patternHeight / ySpacing + 1; iY++) {
           canvas.drawCircle(
             Offset(xStart + xSpacing * iX, yStart + ySpacing * iY),
             lineWidth / 2,
@@ -89,6 +184,20 @@ class PencilDecoration {
           );
         }
       }
+    }
+
+    // Finally draw the frame
+    if (hasFrame) {
+      canvas.drawLine(
+        Offset(xStart + patternWidth, yStart),
+        Offset(xStart + patternWidth, yStart + patternHeight),
+        paint,
+      );
+      canvas.drawLine(
+        Offset(xStart , yStart+ patternHeight),
+        Offset(xStart + patternWidth, yStart + patternHeight),
+        paint,
+      );
     }
   }
 }